Repositioning all module libraries
In the following commit, there was a complete reorganization on the libraries, this was meant to clean the main window in the repository a little bit. The code was tested in uVision and it does compiles without any warnings whatsoever (not tested in actual platform).
Missing to reorganize: pidlib.c because that library adds spaghetti like code.

Still not fixed module dependencies. (e.g.  pwmlib.c uses uartlib)