MPU6050 first version working
Added 3 functions in the library of MPU6050. The first one is to initialize the module, this comprehends the i2c module and a set of registers to initialize the i2c master module. The second one is to perform a single write to a register in the MPU6050. The third one is to perform a single read to a register in the MPU. Planning to add on the next commit a function to read various registers at once.

Added to the main the implementation of the mpu6050.c module and the uart configuration to read the device through realterm.