Working remake of main program.
Added the module of the quadrature encoders, this allows to read the position directly into the encoder.

Added the motor module in this you can configure 2 different motors and write a positive or negative duty cycle to control the speed and direction of any of the two motors.

QEI and PWM working, missing the UART functionality and the pid library.