Final version previous to the MPU6050 branch merge.
Changed the libraries of the Timer 0, Timer 1 and UART to the barr standard and added functions to check the status of the interrupts.

Removed library dependencies so each module can work on its own.

The program now works with or without UART and has a debug feature to check the responsivness of the encoder positions and the PID calculations.

Everything seems to be working.