Working version of motor plus mpu6050
Added a timer2 library because the Tiva C is reaching it's limits in terms of memory and the motors crashed. Tried to fix this error setting the MPU to 15Hz the PID to 100Hz and the Reference switch to 6Hz.

There seems to be no errors right now but further testing is required to check the stability of the motors.